+++ RELEASENOTES.md
@@ -1,3 +1,4 @@
 # 1.1.1
 
 * Added `Nulable<T>` and `Enum` type column tests to support that features.
+* `OnQueryExecuting` method added to `SelectOperation<T>`